git-svn-id: https://svn.code.sf.net/p/opendcl/code/Trunk@92 07a8c159-f786-4120-a571-e88efff3e5ae

+++ Runtime/Controls/ProgressTimeToComplete.cpp
@@ -220,7 +220,7 @@ CString TProgressTimeToComplete::GetRemainingText(double lfPercent, double lfSec
 		double dPercentage = double(nPos) / double(nUpper);
 		dPercentage *= 100;
 
-		str.Format(_T("%d"),(long)dPercentage);
+		str.Format(_T("%d"),(int)dPercentage);
 
 		str += sPercentage;
